Merge diegomura/react-pdf#2895: Take padding into consideration when deciding to break on minPresenceAhead
Fixes diegomura/react-pdf#2884 - Changes splitPage to pass the box paddingTop through to shouldBreak to be used when deciding if breaking a node will improve its presence. Without this shouldBreak can return true for a node that's already as high as it can go on the page, causing an infinite loop.